move filter_map_next to its own module
